# src/ResourceManager/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ResourceManager | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/weatherforecast": {
      "get": {
        "tags": [
          "ResourceManager"
        ],
        "operationId": "GetWeatherForecast",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/resource-event-outbox": {
      "post": {
        "tags": [
          "ResourceManager"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/resource-event": {
      "post": {
        "tags": [
          "ResourceManager"
        ],
        "operationId": "resource-event",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceProcessorEvent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Resource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "resourceId": {
            "type": "string"
          },
          "resourceName": {
            "type": "string"
          },
          "resourceType": {
            "type": "string"
          },
          "accountId": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "availabilityZone": {
            "type": "string"
          },
          "configuration": {
            "type": "string"
          },
          "resourceCreationTime": {
            "type": "string",
            "format": "date-time"
          },
          "resourceHash": {
            "type": "string"
          },
          "tags": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "resource_url": {
            "type": "string"
          }
        }
      },
      "ResourceProcessorEvent": {
        "required": [
          "event_id",
          "provider"
        ],
        "type": "object",
        "properties": {
          "event_id": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "insert_resources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Resource"
            }
          },
          "delete_resources": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "update_resources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/#/properties/insert_resources/items"
            }
          },
          "insert_ships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourceRelationship"
            }
          },
          "delete_ships": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ResourceRelationship": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "relation": {
            "type": "string"
          },
          "sourceId": {
            "type": "string"
          },
          "targetId": {
            "type": "string"
          }
        }
      },
      "WeatherForecast": {
        "required": [
          "date",
          "temperatureC",
          "summary"
        ],
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "temperatureC": {
            "type": "integer",
            "format": "int32"
          },
          "summary": {
            "type": "string",
            "nullable": true
          },
          "temperatureF": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "ResourceManager"
    }
  ]
}